remove old client folder, rename **phaser-demo**
