+++ client/index.html
@@ -1,56 +1,26 @@
 <!DOCTYPE html>
-<html>
+<html lang="en">
 <head>
-    <script src="//cdn.jsdelivr.net/npm/phaser@3.80.1/dist/phaser.min.js"></script>
-    <script src="https://cdn.jsdelivr.net/npm/colyseus.js@0.16.15/dist/colyseus.js"></script>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Phaser Minimal Demo</title>
     <style>
-        /* ... (existing styles) ... */
-        #lobby-info { background-color: rgba(0,50,100,0.7); padding: 15px; margin-bottom: 10px; border-radius: 5px;}
-        #player-list { list-style: none; padding: 0; max-height: 100px; overflow-y: auto; }
-        #player-list li { margin-bottom: 3px; }
-        .ready { color: lightgreen; font-weight: bold; }
-        .not-ready { color: orange; }
-        #instructions-panel { background-color: rgba(50, 50, 0, 0.7); padding: 15px; margin-bottom: 10px; border-radius: 5px; }
-        #admin-controls button { background-color: #ffcc00; border: none; color: black;}
-        #readyButton { background-color: #4CAF50; color: white; padding: 10px 15px; border: none; border-radius: 4px; cursor: pointer; font-size: 16px; }
-        #readyButton.notready { background-color: #ff9800; }
+        body {
+            margin: 0;
+            padding: 0;
+            display: flex;
+            justify-content: center;
+            align-items: center;
+            min-height: 100vh;
+            background-color: #222;
+        }
+        #game-container {
+            box-shadow: 0 0 20px rgba(0, 0, 0, 0.5);
+        }
     </style>
-    <link rel="icon" href="favicon.ico">
 </head>
 <body>
-    <div id="phaser-game"></div>
-
-    <div id="ui-overlay" style="display: none;">
-        <div id="lobby-info">
-            <h4>Lobby</h4>
-            <p id="lobby-message">Connecting...</p>
-            <ul id="player-list"></ul>
-            <button id="readyButton" class="notready">Ready</button>
-        </div>
-
-         <div id="instructions-panel" style="display: none;">
-              <h4>Instructions</h4>
-              <p id="instructions-text"></p>
-         </div>
-
-
-        <div id="gameplay-info" style="display: none;">
-             <div>Status: <span id="game-status">Starting...</span></div>
-             <div>Round: <span id="round-number">N/A</span></div>
-             <div id="role-info">Role: <span id="player-role">Waiting...</span></div>
-             <div id="target-info" style="display: none;">Target Zone: <span id="target-zone-text">-</span></div>
-         </div>
-
-    </div>
-
-    <div id="timer" style="display: none;">00:00</div>
-     <div id="score" style="display: none;">Score: 0</div>
-    <div id="final-message" style="display: none;">
-        </div>
-    <div id="admin-controls" style="display: none;"> <button id="startGameBtn">Start Experiment</button>
-         <p>Status: <span id="serverStateDisplay">Unknown</span></p>
-    </div>
-
-    <script type="module" src="main.ts"></script>
+    <div id="game-container"></div>
+    <script type="module" src="./src/main.ts"></script>
 </body>
 </html>
