+++ client/package.json
@@ -1,33 +1,20 @@
 {
-  "name": "flocker-phaser",
+  "name": "phaser-minimal-demo",
   "version": "1.0.0",
-  "description": "Online experiment for collective decision making",
+  "description": "Minimal Phaser demo with Colyseus integration",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "parcel serve index.html",
+    "start": "parcel index.html",
     "build": "parcel build index.html"
   },
-  "author": "gus-bradley",
+  "author": "",
   "license": "ISC",
   "devDependencies": {
-    "assert": "^2.1.0",
     "buffer": "^6.0.3",
-    "crypto-browserify": "^3.12.1",
-    "events": "^3.3.0",
-    "os-browserify": "^0.3.0",
     "parcel": "^2.16.0",
-    "path-browserify": "^1.0.1",
     "process": "^0.11.10",
-    "stream-browserify": "^3.0.0",
-    "stream-http": "^3.2.0",
-    "tsx": "^4.20.6",
-    "typescript": "^5.9.3",
-    "url": "^0.11.4",
-    "util": "^0.12.5",
-    "vm-browserify": "^1.1.2"
+    "typescript": "^5.9.3"
   },
   "dependencies": {
-    "colyseus": "^0.16.5",
     "colyseus.js": "^0.16.22",
     "phaser": "^3.90.0"
   }
