+++ client/src/scenes/GameScene.ts
@@ -1,492 +1,467 @@
+import Phaser, { NONE } from 'phaser';
+import * as Colyseus from 'colyseus.js';
+import { Player, Zone, RoomState, Phase } from '../../../server/src/rooms/schema/experimentSchema'
 
-import Phaser from "phaser";
+import { BACKEND_HTTP_URL, BACKEND_URL } from "../backend";
+import { config } from '../config';
 
 export default class GameScene extends Phaser.Scene {
+    private playerSpeed = config.player.speed;
+    private lastPosition = { x: 0, y: 0 };
+    private lastSentPosition = { x: 0, y: 0 };
+    private lastPositionUpdateTime = 0;
+    private lastCollisionPositionUpdateTime = 0; // Separate timer for collision updates
+    private positionUpdateThreshold = config.network.positionUpdateThreshold;
+    private positionUpdateInterval = config.network.positionUpdateInterval;
+    private collisionUpdateInterval = config.network.collisionUpdateInterval;
+
+    // Colyseus properties
+    room: Colyseus.Room<RoomState>;
+    playerEntities: { [sessionId: string]: Phaser.GameObjects.Container } = {};
+	zoneEntities: { [zoneId: string]: Phaser.GameObjects.Arc } = {};
+    cursorKeys: Phaser.Types.Input.Keyboard.CursorKeys;
+    keys: { [key: string]: Phaser.Input.Keyboard.Key };
+    playersGroup: Phaser.Physics.Arcade.Group;
+    debugText: Phaser.GameObjects.Text;
+    timerText: Phaser.GameObjects.Text;
+    readyButton: Phaser.GameObjects.Text;
+    readyCount: Phaser.GameObjects.Text;
+    isHost: boolean = false;
+    // currentPlayer: Phaser.GameObjects.Container;
+
     constructor() {
-        super({ key: 'GameScene' });
-        this.playerSprites = {};
-        this.zoneSprites = {};
-        this.selfId = null;
-        this.playerSpeed = 150;
-        this.room = null; // Colyseus room instance
-        this.client = null; // Colyseus client instance
-
-        // UI Element references
-         this.uiOverlay = null;
-         this.lobbyInfoPanel = null;
-         this.lobbyMessageDisplay = null;
-         this.playerListDisplay = null;
-         this.readyButton = null;
-         this.instructionsPanel = null;
-         this.instructionsTextDisplay = null;
-         this.gameplayInfoPanel = null;
-         this.gameStatusDisplay = null;
-         this.roundDisplay = null;
-         this.roleDisplay = null;
-         this.targetInfoPanel = null;
-         this.targetZoneText = null;
-         this.timerDisplay = null;
-         this.scoreDisplay = null;
-         this.finalMessagePanel = null;
-         this.adminControlsPanel = null;
-         this.serverStateDisplay = null; // For admin
-
-        this.emoteSprites = {};
-        this.emoteTimers = {};
-
-        this.isReady = false; // Track client's own ready state
-         this.isAdmin = false; // Track if this client is the admin
+        super({ key: "experiment" });
     }
 
-    preload() {
-        // Preload scene should still handle assets
-        console.log("GameScene: Preload/Create sequence starting...");
-    }
+    // preload() {
+    // }
 
+    async create() {
+        this.cursorKeys = this.input.keyboard!.createCursorKeys();
 
-    create() {
-        console.log("GameScene: Create");
-        this.physics.world.setBounds(0, 0, this.game.config.width, this.game.config.height);
+        // Create keys object that can recognize number keys 1-4
+        this.keys = this.input.keyboard!.addKeys({
+            'ONE': Phaser.Input.Keyboard.KeyCodes.ONE,
+            'TWO': Phaser.Input.Keyboard.KeyCodes.TWO,
+            'THREE': Phaser.Input.Keyboard.KeyCodes.THREE,
+            'FOUR': Phaser.Input.Keyboard.KeyCodes.FOUR,
+        }) as { [key: string]: Phaser.Input.Keyboard.Key };
 
-        // Get UI Element References
-         this.uiOverlay = document.getElementById('ui-overlay');
-         this.lobbyInfoPanel = document.getElementById('lobby-info');
-         this.lobbyMessageDisplay = document.getElementById('lobby-message');
-         this.playerListDisplay = document.getElementById('player-list');
-         this.readyButton = document.getElementById('readyButton');
-         this.instructionsPanel = document.getElementById('instructions-panel');
-         this.instructionsTextDisplay = document.getElementById('instructions-text');
-         this.gameplayInfoPanel = document.getElementById('gameplay-info');
-         this.gameStatusDisplay = document.getElementById('game-status');
-         this.roundDisplay = document.getElementById('round-number');
-         this.roleDisplay = document.getElementById('player-role');
-         this.targetInfoPanel = document.getElementById('target-info');
-         this.targetZoneText = document.getElementById('target-zone-text');
-         this.timerDisplay = document.getElementById('timer');
-         this.scoreDisplay = document.getElementById('score');
-         this.finalMessagePanel = document.getElementById('final-message');
-         this.adminControlsPanel = document.getElementById('admin-controls');
-         this.serverStateDisplay = document.getElementById('serverStateDisplay'); // Admin
+        this.cameras.main.setBackgroundColor(0xFFFFFF);
 
-        // --- Connect to Colyseus ---
-        this.connect();
+        // connect with the room
+        await this.connect();
 
-        // --- Input Handling --- (Remains similar)
-        this.input.on('pointerdown', (pointer) => {
-            if (this.room && this.selfId && this.playerSprites[this.selfId] && this.room.state.phase === 'round_active') {
-                const targetPos = { x: pointer.worldX, y: pointer.worldY };
-                this.playerSprites[this.selfId].setData('targetPos', targetPos);
-                // Send target to server via Colyseus message
-                this.room.send('playerMovement', targetPos);
+        // Create physics group for players
+        this.playersGroup = this.physics.add.group();
+
+        // Add collider to prevent players from overlapping
+        this.physics.add.collider(this.playersGroup, this.playersGroup, () => {
+            // Check if room exists before accessing it
+            if (!this.room) return;
+
+            const currentTime = this.time.now;
+            const timeSinceLastCollisionUpdate = currentTime - this.lastCollisionPositionUpdateTime;
+
+            // Only send position update if enough time has passed
+            if (timeSinceLastCollisionUpdate >= this.collisionUpdateInterval) {
+                const playerEntity = this.playerEntities[this.room.sessionId];
+                if (playerEntity) {
+                    this.room.send("position", { "x": playerEntity.x, "y": playerEntity.y });
+                    this.lastCollisionPositionUpdateTime = currentTime;
+                }
             }
         });
 
-        this.input.keyboard.on('keydown-ONE', () => this.sendEmote('1'));
-        this.input.keyboard.on('keydown-TWO', () => this.sendEmote('2'));
-        this.input.keyboard.on('keydown-THREE', () => this.sendEmote('3'));
-        this.input.keyboard.on('keydown-FOUR', () => this.sendEmote('4'));
+        const $ = Colyseus.getStateCallbacks(this.room);
 
-        // Ready Button Logic
-        this.readyButton.addEventListener('click', () => {
-             if (this.room && this.room.state.phase === 'lobby') {
-                 this.isReady = !this.isReady;
-                 this.room.send('setReady', { isReady: this.isReady });
-                 this.updateReadyButtonVisuals();
-             }
+        // Handle new zones added after we join
+        $(this.room.state).zones.onAdd((zone, zoneId) => {
+            console.log("New zone added:", zoneId, zone);
+            // Convert hex color string to number (e.g., "#FF0000" -> 0xFF0000)
+            const colorNumber = parseInt(zone.color.replace('#', ''), 16);
+
+            // Create an arc (circle) for this zone
+            const entity = this.add.arc(zone.x, zone.y, zone.radius, 0, 360, false, colorNumber, config.zones.opacity);
+            entity.setDepth(-1);
+            if (this.room.state.targetZone === zone.id && this.room.state.players.get(this.room.sessionId)?.informed) {
+                entity.setStrokeStyle(config.zones.targetWidth, config.zones.targetColor);
+            }
+            this.zoneEntities[zone.id] = entity;
         });
 
-        // Admin Button Logic (if admin)
-         const startGameBtn = document.getElementById('startGameBtn');
-         if (startGameBtn) {
-             startGameBtn.addEventListener('click', () => {
-                 if (this.room && this.isAdmin) {
-                     this.room.send('admin_startGame');
-                 }
-             });
-         }
+        // Listen for targetZone changes to update stroke
+        $(this.room.state).listen('targetZone', (value) => {
+            console.log('Target zone changed:', value);
 
-    } // End of create()
+            // Clear stroke from all zones
+            Object.keys(this.zoneEntities).forEach((zoneId) => {
+                if (zoneId === value.toString() && this.room.state.players.get(this.room.sessionId)?.informed) {
+                    this.zoneEntities[zoneId].setStrokeStyle(config.zones.targetWidth, config.zones.targetColor);
+                }
+                else {
+                    this.zoneEntities[zoneId].setStrokeStyle(0);
+                }
+            });
+        });
 
-    updateReadyButtonVisuals() {
-         if (this.isReady) {
-             this.readyButton.textContent = "Ready!";
-             this.readyButton.classList.remove("notready");
-             this.readyButton.classList.add("ready");
-         } else {
-             this.readyButton.textContent = "Ready?";
-             this.readyButton.classList.remove("ready");
-             this.readyButton.classList.add("notready");
-         }
-     }
-
-     async connect() {
-        try {
-            const protocol = window.location.protocol === 'https:' ? 'wss' : 'ws';
-            const endpoint = `${protocol}://${window.location.host}`;
+        // Handle new players added after we join
+        $(this.room.state).players.onAdd((player, sessionId) => {
+            console.log("New player added:", sessionId, player);
             
-            console.log(`Connecting to Colyseus at ${endpoint}`);
-            this.client = new Colyseus.Client(endpoint);
-    
-            console.log("Attempting to join 'experiment_room'...");
-            this.room = await this.client.joinOrCreate('experiment_room');
-            this.selfId = this.room.sessionId;
-            console.log("Successfully joined room!", this.room.name, "My Session ID:", this.selfId);
-            this.uiOverlay.style.display = 'block';
-    
-            this.registerRoomEvents();
-        } catch (e) {
-            console.error("Join Error:", e);
-            this.lobbyMessageDisplay.textContent = `Failed to connect: ${e.message}`;
-        }
-    }
+            // Convert hex color string to number (e.g., "#FF0000" -> 0xFF0000)
+            const colorNumber = parseInt(player.color.replace('#', ''), 16);
 
-    registerRoomEvents() {
-        if (!this.room) return;
-    
-        // Wait for the state to be fully initialized
-        this.room.onStateChange.once((state) => {
-            if (!state.players) {
-                console.error("State synchronization failed: 'players' is undefined.");
-                return;
+            // Create an arc (circle) for the player
+            const circle = this.add.arc(0, 0, player.radius, 0, 360, false, colorNumber, 1);
+
+            // Create text to display the emote in the middle of the circle
+            const text = this.add.text(0, 0, player.emote.toString(), {
+                fontSize: config.ui.playerText.fontSize,
+                color: player.textColor,
+                fontStyle: config.ui.playerText.fontStyle
+            });
+            text.setOrigin(0.5, 0.5); // Center the text
+
+            // Create a container to group the circle and text
+            const container = this.add.container(player.x, player.y, [circle, text]);
+
+            // Enable physics on the container
+            this.physics.world.enable(container);
+            const body = container.body as Phaser.Physics.Arcade.Body;
+            body.setCircle(player.radius);
+            body.setCollideWorldBounds(true);
+
+
+            this.playerEntities[sessionId] = container;
+
+            // if (sessionId === this.room.sessionId) {
+            //     this.currentPlayer = container;
+            // }
+
+            // Add to players group for collision detection
+            this.playersGroup.add(container);
+
+            // Update ready count when a player is added
+            // this.updateReadyCount();
+
+            // Listen for ready status changes
+            // $(player).listen('ready', () => {
+            //     this.updateReadyCount();
+            // });
+
+            // listening for server updates
+            $(player).onChange(() => {
+
+                container.x = player.x;
+                container.y = player.y;
+
+                // Update the emote text for all players
+                const textElement = container.getAt(1) as Phaser.GameObjects.Text;
+                textElement.setText(player.emote.toString());
+
+                if (player === this.getCurrentPlayer()) {
+                    Object.keys(this.zoneEntities).forEach((zoneId) => {
+                        if (zoneId === this.room.state.targetZone.toString() && this.room.state.players.get(this.room.sessionId)?.informed) {
+                            this.zoneEntities[zoneId].setStrokeStyle(config.zones.targetWidth, config.zones.targetColor);
+                        }
+                        else {
+                            this.zoneEntities[zoneId].setStrokeStyle(0);
+                        }
+                    });
+                }
+
+            });
+        });
+
+        // remove local reference when entity is removed from the server
+        $(this.room.state).players.onRemove((_player, sessionId) => {
+            console.log("Player removed:", sessionId);
+            const entity = this.playerEntities[sessionId];
+            if (entity) {
+                this.playersGroup.remove(entity, true, true); // remove from group and destroy
+                delete this.playerEntities[sessionId]
             }
-    
-            console.log("State synchronized:", state);
-    
-            // --- Player Add/Remove/Change Listeners ---
-            state.players.onAdd((player, sessionId) => {
-                console.log("Player added:", sessionId, player);
-                if (!this.playerSprites[sessionId]) {
-                    this.addPlayerSprite(player, sessionId);
+            // Update ready count when a player is removed
+            // this.updateReadyCount();
+        });
+
+        // add debugging text
+        const debugText = this.add
+            .text(config.ui.debug.position.x, config.ui.debug.position.y, "debug text here")
+            .setStyle({ color: config.ui.debug.color, fontSize: config.ui.debug.fontSize })
+        this.debugText = debugText
+
+        // add round timer text
+        const timerText = this.add
+            .text(config.ui.timer.position.x, config.ui.timer.position.y, config.ui.timer.prefix)
+            .setStyle({ color: config.ui.timer.color, fontSize: config.ui.timer.fontSize })
+        this.timerText = timerText
+
+        // // Add start button (initially hidden)
+        this.readyButton = this.add
+            .text(400, 300, 'Ready', {
+                fontSize: '32px',
+                color: '#000000ff',
+                backgroundColor: '#0717ff6e',
+                padding: { x: 20, y: 10 }
+            })
+            .setOrigin(0.5)
+            .setInteractive({ useHandCursor: true })
+            .setVisible(false)
+            .on('pointerdown', () => {
+                const currentPlayer = this.getCurrentPlayer();
+                if (currentPlayer && !currentPlayer.ready) {
+                    this.room.send('ready');
+                    this.readyButton.setText("Ready ✓");
+                    this.readyButton.setStyle({ backgroundColor: '#00ff006e'});
+                }
+            })
+            .setDepth(2);
+        
+        // this.readyCount = this.add
+        //     .text(400, 200, `${Array.from(this.room.state.players.values()).filter(p => p.ready).length}/${this.room.state.players.size} ready`, {
+        //         fontSize: '32px',
+        //         color: '#000000ff',
+        //         padding: {x: 20, y:10}
+        //     })
+        //     .setOrigin(0.5)
+        //     .setVisible(false)
+        //     .setInteractive(false)
+        //     .setDepth(2)
+
+        $(this.room.state).listen('phase', (value) => {
+            const currentPlayer = this.getCurrentPlayer();
+
+            if (value === Phase.WAITING) {
+                this.readyButton.setVisible(true);
+                // this.readyCount.setVisible(true);
+                // this.updateReadyCount();
+                if (currentPlayer && !currentPlayer.ready) {
+                    this.readyButton.setText('Ready');
+                    this.readyButton.setStyle({backgroundColor: '#0717ff6e'});
+                }
+            } else {
+                this.readyButton.setVisible(false);
+                this.readyCount.setVisible(false);
+            }
+        })
+
+        const currentPlayer = this.getCurrentPlayer();
+        if (currentPlayer) {
+            $(currentPlayer).listen('ready', (value) => {
+                if (value) {
+                    this.readyButton.setText("Ready ✓");
+                    this.readyButton.setStyle({backgroundColor: '#00ff006e'});
+                    this.readyButton.disableInteractive();
                 } else {
-                    this.updatePlayerSprite(this.playerSprites[sessionId], player);
+                    this.readyButton.setText('Ready');
+                    this.readyButton.setStyle({backgroundColor: '#0717ff6e'});
+                    this.readyButton.setInteractive({useHandCursor: true});
                 }
-                if (state.phase === 'lobby') this.updatePlayerList(state.players);
+
             });
-    
-            state.players.onRemove((player, sessionId) => {
-                console.log("Player removed:", sessionId);
-                this.removePlayerSprite(sessionId);
-                if (state.phase === 'lobby') this.updatePlayerList(state.players);
+        }
+        
+            // .on('pointerover', () => {
+            //     if (this.isHost) {
+            //         this.startButton.setStyle({ color: '#ffffff' });
+            //     }
+            // })
+            // .on('pointerout', () => {
+            //     this.startButton.setStyle({ color: '#00ff00' });
+            // });
+
+        // // Listen for roundActive state changes
+        // $(this.room.state).listen('roundActive', (value) => {
+        //     console.log('Round active changed:', value);
+        //     // Hide button when round is active
+        //     if (this.isHost) {
+        //         this.startButton.setVisible(!value);
+        //     }
+        // });
+
+        // // Listen for changes to current player's host status
+        // const currentPlayer = this.room.state.players.get(this.room.sessionId);
+        // if (currentPlayer) {
+        //     this.isHost = currentPlayer.host;
+        //     this.startButton.setVisible(this.isHost && !this.room.state.roundActive);
+
+        //     $(currentPlayer).listen('host', (value) => {
+        //         console.log('Host status changed:', value);
+        //         this.isHost = value;
+        //         this.startButton.setVisible(this.isHost && !this.room.state.roundActive);
+        //     });
+        // }
+    }
+
+    async connect() {
+
+        const client = new Colyseus.Client(BACKEND_URL);
+
+        try {
+
+            this.room = await client.joinOrCreate("ExperimentRoom", {})
+
+        } catch (e) {
+
+            console.error("Connection error:", e);
+        }
+
+    }
+
+    update(time: number, delta: number): void {
+        // skip loop if not connected with room yet.
+        if (!this.room) {
+            return;
+        }
+
+        // calculate movement delta based on arrow keys
+        let dx = 0;
+        let dy = 0;
+        let emote = "";
+
+        if (this.cursorKeys.left.isDown) {
+            dx -= 1;
+        }
+        if (this.cursorKeys.right.isDown) {
+            dx += 1;
+        }
+        if (this.cursorKeys.up.isDown) {
+            dy -= 1;
+        }
+        if (this.cursorKeys.down.isDown) {
+            dy += 1;
+        }
+        if (this.keys.ONE.isDown) {
+            emote = config.emotes.ONE;
+        }
+        if (this.keys.TWO.isDown) {
+            emote = config.emotes.TWO;
+        }
+        if (this.keys.THREE.isDown) {
+            emote = config.emotes.THREE;
+        }
+        if (this.keys.FOUR.isDown) {
+            emote = config.emotes.FOUR;
+        }
+
+        // only send movement if there's a delta and round is active
+        if (dx !== 0 || dy !== 0) {
+            this.room.send("move", { "x": dx, "y": dy });
+        }
+
+        // Check if selected emote is different from current character emote
+        if (!this.room.state.players) {
+            return;
+        }
+        const currentPlayer = this.getCurrentPlayer();
+        
+        if (emote !== "" && currentPlayer && emote !== currentPlayer.emote) {
+            this.room.send("emote", { "emote": emote });
+        }
+
+        // Sync position with server if it differs (for collision-based movements)
+        // if (currentPlayer) {
+        //     const playerEntity = this.playerEntities[this.room.sessionId];
+
+        //     if (playerEntity) {
+        //         const actualX = Math.round(playerEntity.x);
+        //         const actualY = Math.round(playerEntity.y);
+        //         const serverX = Math.round(currentPlayer.x);
+        //         const serverY = Math.round(currentPlayer.y);
+
+        //         // Calculate distance between actual position and server position
+        //         const dx = Math.abs(actualX - serverX);
+        //         const dy = Math.abs(actualY - serverY);
+
+        //         if (dy > this.positionUpdateThreshold || dx > this.positionUpdateThreshold) {
+        //             this.room.send("position", { "x": actualX, "y": actualY });
+        //             console.log('Sent position message: x =' + actualX + " y = " + actualY)
+        //             this.lastSentPosition = { x: actualX, y: actualY };
+        //             this.lastPositionUpdateTime = time;
+        //         }
+        //     }
+        // }
+
+        // update debug text
+        if (currentPlayer) {
+            this.debugText.text = this.generateDebugText({
+                session_id: this.room.sessionId,
+                zone_id: currentPlayer.zone,
+                room_id: this.room.roomId,
+                target_zone: this.room.state.targetZone,
+                player_points: currentPlayer.points,
+                host: currentPlayer.host,
+                phase: this.room.state.phase,
+                informed: currentPlayer.informed,
+                ready: currentPlayer.ready,
             });
-    
-            state.players.onChange((player, sessionId) => {
-                const sprite = this.playerSprites[sessionId];
-                if (sprite) {
-                    this.updatePlayerSprite(sprite, player);
-                }
-                if (state.phase === 'lobby') this.updatePlayerList(state.players);
-            });
-        });
-    
-        // Handle other room events (e.g., messages, errors)
-        this.room.onMessage("isAdmin", (isAdmin) => {
-            console.log(`Received isAdmin status: ${isAdmin}`);
-            this.isAdmin = isAdmin;
-            this.adminControlsPanel.style.display = isAdmin ? 'block' : 'none';
-        });
-    
-        this.room.onMessage("error", (message) => {
-            console.error("Server Error Message:", message);
-            alert(`Server Error: ${message}`);
-        });
-    
-        // Add other event handlers as needed
-    }
-
-     updateUIPhase(phase, state) {
-         // Hide all panels by default
-         this.lobbyInfoPanel.style.display = 'none';
-         this.instructionsPanel.style.display = 'none';
-         this.gameplayInfoPanel.style.display = 'none';
-          this.finalMessagePanel.style.display = 'none'; // Ensure final message hidden unless experiment end
-          this.scoreDisplay.style.display = 'none'; // Hide score unless active/end
-
-
-         // Show relevant panel based on phase
-         switch (phase) {
-             case 'lobby':
-                 this.lobbyInfoPanel.style.display = 'block';
-                 this.lobbyMessageDisplay.textContent = state.lobbyMessage || "Waiting for players...";
-                 this.updatePlayerList(state.players); // Ensure list is updated
-                 break;
-             case 'instructions':
-                 this.instructionsPanel.style.display = 'block';
-                 this.instructionsTextDisplay.textContent = state.instructionText;
-                 break;
-             case 'round_starting':
-                 this.gameplayInfoPanel.style.display = 'block';
-                 this.gameStatusDisplay.textContent = `Starting Round ${state.roundNumber}...`;
-                 this.roundDisplay.textContent = state.roundNumber;
-                 // Role/Target info updated via player state change
-                 break;
-             case 'round_active':
-                 this.gameplayInfoPanel.style.display = 'block';
-                 this.gameStatusDisplay.textContent = `Round ${state.roundNumber} Active`;
-                 this.roundDisplay.textContent = state.roundNumber;
-                 this.scoreDisplay.style.display = 'block';
-                 // Update own role/target info if changed
-                 this.updateSelfInfo(state.players.get(this.selfId), state.targetZone);
-                 break;
-             case 'round_end':
-                 this.gameplayInfoPanel.style.display = 'block';
-                 this.gameStatusDisplay.textContent = `Round ${state.roundNumber} Ended`;
-                 this.roundDisplay.textContent = state.roundNumber;
-                 this.scoreDisplay.style.display = 'block';
-                 this.targetInfoPanel.style.display = 'none'; // Hide target zone info
-                  // Stop player movement visually
-                 for (const playerId in this.playerSprites) {
-                     this.playerSprites[playerId].setData('targetPos', null);
-                      if(this.playerSprites[playerId].body) { // Check body exists
-                           this.playerSprites[playerId].body.velocity.reset();
-                      }
-                 }
-                 break;
-             case 'experiment_end':
-                  // Handled by experimentEndInfo message typically
-                  // If relying solely on state:
-                 // this.finalMessagePanel.style.display = 'block';
-                 // Update final message content here from state if needed
-                 break;
-             default:
-                  this.lobbyMessageDisplay.textContent = `Unknown phase: ${phase}`;
-                 this.lobbyInfoPanel.style.display = 'block';
-         }
-     }
-
-    updateSelfInfo(playerState, targetZone) {
-        if (playerState) {
-             this.roleDisplay.textContent = playerState.role;
-             if (playerState.role === 'informed' && targetZone) {
-                 this.targetZoneText.textContent = targetZone;
-                 this.targetInfoPanel.style.display = 'block';
-             } else {
-                 this.targetInfoPanel.style.display = 'none';
-             }
-        } else {
-            // Own player state not found? Handle error/wait.
-             this.roleDisplay.textContent = 'Unknown';
-             this.targetInfoPanel.style.display = 'none';
-        }
-    }
-
-     updatePlayerList(playersMap) {
-        if (!this.playerListDisplay) return;
-         this.playerListDisplay.innerHTML = ''; // Clear previous list
-         playersMap.forEach((player, sessionId) => {
-             const li = document.createElement('li');
-             const readyStatus = player.ready ? '<span class="ready">(Ready)</span>' : '<span class="not-ready">(Not Ready)</span>';
-             li.innerHTML = `${player.name || 'Player'} ${sessionId === this.selfId ? '(You)' : ''} ${readyStatus}`;
-             this.playerListDisplay.appendChild(li);
-         });
-     }
-
-
-    addPlayerSprite(playerData, sessionId) {
-         // console.log(`Attempting to add sprite for ${sessionId}`);
-         let playerSprite = this.physics.add.sprite(playerData.x, playerData.y, 'player');
-         playerSprite.setCollideWorldBounds(true);
-         playerSprite.setCircle(playerSprite.width / 2);
-         playerSprite.setTint(playerData.color);
-         playerSprite.setData('id', sessionId);
-         playerSprite.setData('targetPos', null);
-         playerSprite.setData('initialColor', playerData.color); // Store initial color
-         this.playerSprites[sessionId] = playerSprite;
-
-         let emoteSprite = this.add.sprite(playerData.x, playerData.y - 20, 'emotes');
-         emoteSprite.setVisible(false);
-         emoteSprite.setOrigin(0.5, 1);
-         // emoteSprite.setTint(playerData.emoteColor); // Apply emote color?
-         this.emoteSprites[sessionId] = emoteSprite;
-
-          // Update emote state if initially provided
-         if (playerData.currentEmote) {
-             this.showEmote(sessionId, playerData.currentEmote);
-         }
-         console.log(`Added sprite for ${sessionId}`);
-     }
-
-    removePlayerSprite(sessionId) {
-        if (this.playerSprites[sessionId]) {
-            this.playerSprites[sessionId].destroy();
-            delete this.playerSprites[sessionId];
-        }
-         if (this.emoteSprites[sessionId]) {
-            this.emoteSprites[sessionId].destroy();
-            delete this.emoteSprites[sessionId];
-         }
-         if (this.emoteTimers[sessionId]) {
-            clearTimeout(this.emoteTimers[sessionId]);
-            delete this.emoteTimers[sessionId];
-         }
-         console.log(`Removed sprite for ${sessionId}`);
-    }
-
-     updatePlayerSprite(sprite, playerData) {
-        if (!sprite || !playerData) return;
-
-        // Smooth movement for remote players? (or just teleport)
-         if (sprite.getData('id') !== this.selfId) {
-             // Option 1: Teleport
-             // sprite.setPosition(playerData.x, playerData.y);
-             // Option 2: Smooth tweening (more complex, needs target handling)
-              this.physics.moveTo(sprite, playerData.x, playerData.y, this.playerSpeed * 0.5); // Move towards state pos slowly?
-
-         }
-
-         // Update other properties
-        sprite.setTint(playerData.color);
-         sprite.setData('initialColor', playerData.color); // Update stored color
-
-
-        // Update emote state from server state
-        if (playerData.currentEmote) {
-             this.showEmote(sprite.getData('id'), playerData.currentEmote);
-        } else {
-            this.hideEmote(sprite.getData('id'));
         }
 
-         // Update role/target info if it's the self player
-         if (sprite.getData('id') === this.selfId) {
-             this.updateSelfInfo(playerData, this.room.state.targetZone);
-         }
+        this.timerText.text = "" + this.room.state.roundTime;
+
+        //         // Check if enough time has passed since last update
+        //         const timeSinceLastUpdate = time - this.lastPositionUpdateTime;
+
+        //         // Send position update if:
+        //         // 1. Position differs by more than threshold, AND
+        //         // 2. Enough time has passed since last update, AND
+        //         // 3. Position has actually changed since last sent position
+        //         // if ((dx > this.positionUpdateThreshold || dy > this.positionUpdateThreshold) &&
+        //         //     timeSinceLastUpdate >= this.positionUpdateInterval &&
+        //         //     (actualX !== this.lastSentPosition.x || actualY !== this.lastSentPosition.y)) {
+        //         if (timeSinceLastUpdate >= this.positionUpdateInterval) {
+
+        //         }
+        //     }
+        // }
     }
 
-    updateTimerDisplay(endTime) {
-        if (this.timerDisplay) {
-             const remaining = Math.max(0, endTime - this.room.clock.currentTime); // Use Colyseus clock if available and synced
-             // Fallback if clock isn't perfectly synced: use local time approximation based on initial end time
-             // const remaining = Math.max(0, initialEndTimeFromServer - Date.now());
-             const seconds = Math.floor((remaining / 1000) % 60);
-             const minutes = Math.floor((remaining / (1000 * 60)) % 60);
-             this.timerDisplay.textContent = `${minutes.toString().padStart(2, '0')}:${seconds.toString().padStart(2, '0')}`;
-        }
+    private generateDebugText(params: {
+        session_id?: string;
+        zone_id?: number;
+        room_id?: string;
+        target_zone?: number;
+        player_points?: number;
+        host?: boolean;
+        game_status?: boolean;
+        informed?: boolean;
+        phase?: Phase;
+        ready?: boolean;
+    }) {
+        const fields = {
+            'Session ID': params.session_id,
+            'Room ID': params.room_id,
+            'Zone': params.zone_id?.toString(),
+            'Target Zone': params.target_zone?.toString(),
+            'Player Points': params.player_points?.toString(),
+            'Game Status': params.game_status?.toString(),
+            'Host': params.host?.toString(),
+            'Informed': params.informed?.toString(),
+            'Phase': params.phase?.toString(),
+            'Player Ready': params.ready?.toString(),
+        };
+
+        return Object.entries(fields)
+            .map(([label, value]) => `\n${label}: ${value || ''}`)
+            .join('');
     }
 
-    update(time, delta) {
-        if (!this.selfId || !this.room || !this.playerSprites[this.selfId]) return;
-
-         // --- Self Movement ---
-         const selfSprite = this.playerSprites[this.selfId];
-         if (selfSprite.getData('targetPos') && this.room.state.phase === 'round_active') {
-             this.movePlayerTowards(selfSprite, selfSprite.getData('targetPos'), delta);
-         } else if (selfSprite.body && selfSprite.body.velocity.x !== 0 || selfSprite.body.velocity.y !== 0) {
-              selfSprite.body.velocity.reset(); // Ensure stopped if no target
-         }
-
-         // --- Other Player Movement (if using moveTo for smoothing) ---
-         for (const sessionId in this.playerSprites) {
-            if (sessionId !== this.selfId) {
-                const sprite = this.playerSprites[sessionId];
-                const playerState = this.room.state.players.get(sessionId);
-                if (sprite && playerState && sprite.body) {
-                    // Stop smoothing if close enough to the state position
-                    const distance = Phaser.Math.Distance.Between(sprite.x, sprite.y, playerState.x, playerState.y);
-                    if (distance < 4) {
-                        sprite.body.reset(playerState.x, playerState.y); // Snap to final position
-                    }
-                }
-            }
+    private getCurrentPlayer() {
+        if (this.room.state.players) {
+             return this.room.state.players.get(this.room.sessionId);
         }
-
-
-         // --- Update Emote Positions ---
-         this.updateEmotePositions();
-
-         // --- Update Timer --- (Driven by state change now, but could update sub-second here)
-         if (this.room.state.phase === 'round_active' && this.room.state.roundEndTime > 0) {
-              this.updateTimerDisplay(this.room.state.roundEndTime);
-         }
-
-    } // End of update()
-
-    // --- Helper Functions --- (drawZone, movePlayerTowards, sendEmote, show/hideEmote, highlightPlayers, resetHighlighting - adapt as needed)
-
-     drawZone(zoneData) {
-        if (!this.zoneSprites[zoneData.id]) {
-            let zoneSprite = this.add.sprite(zoneData.x, zoneData.y, 'zone');
-            zoneSprite.setAlpha(0.7);
-            zoneSprite.setDisplaySize(zoneData.radius * 2, zoneData.radius * 2);
-            this.add.text(zoneData.x, zoneData.y, zoneData.id, { fontSize: '16px', fill: '#000' }).setOrigin(0.5);
-            this.zoneSprites[zoneData.id] = zoneSprite;
-        }
+        return null;
     }
 
-    movePlayerTowards(playerSprite, targetPos, delta) {
-        const distance = Phaser.Math.Distance.Between(playerSprite.x, playerSprite.y, targetPos.x, targetPos.y);
-        if (distance < 5) {
-             if (playerSprite.body) playerSprite.body.velocity.reset();
-             playerSprite.setData('targetPos', null);
-             // Send final position update? Optional. Colyseus state might lag slightly.
-             // this.room.send("positionUpdate", { x: playerSprite.x, y: playerSprite.y });
-        } else if (playerSprite.body) {
-             const angle = Phaser.Math.Angle.Between(playerSprite.x, playerSprite.y, targetPos.x, targetPos.y);
-             this.physics.velocityFromRotation(angle, this.playerSpeed, playerSprite.body.velocity);
+    // private updateReadyCount() {
+    //     if (!this.readyCount || !this.room || !this.room.state.players) return;
 
-             // --- CRITICAL for Server Accuracy ---
-             // Send frequent position updates while moving
-             const lastSent = playerSprite.getData('lastSentPosTime') || 0;
-             if (this.room.clock.currentTime - lastSent > 100) { // Send every 100ms
-                  // console.log("Sending position update"); // DEBUG
-                  // this.room.send("positionUpdate", { x: playerSprite.x, y: playerSprite.y }); // Need server handler
-                  playerSprite.setData('lastSentPosTime', this.room.clock.currentTime);
-                  // IMPORTANT: Server MUST use these updates to update player.x/y in its state for accurate logging/zone checks
-             }
+    //     const readyPlayers = Array.from(this.room.state.players.values()).filter(player => player.ready).length;
+    //     const totalPlayers = this.room.state.players.size;
 
-        }
-    }
+    //     this.readyCount.setText(`${readyPlayers}/${totalPlayers} ready`);
+    // }
 
-     sendEmote(emoteKey) {
-        if (this.room && this.selfId && (this.room.state.phase === 'round_active' || this.room.state.phase === 'lobby' || this.room.state.phase === 'instructions')) {
-            console.log(`Sending emote: ${emoteKey}`);
-            this.room.send('playerEmote', emoteKey);
-             // Show emote locally immediately for responsiveness
-            const emoteMap = { '1': '!', '2': '+', '3': 'x', '4': '?' };
-            this.showEmote(this.selfId, emoteMap[emoteKey]);
-        }
-    }
-
-     showEmote(playerId, emoteSymbol) {
-         // ... (Implementation similar to previous version, using this.emoteSprites, this.emoteTimers) ...
-         const playerSprite = this.playerSprites[playerId];
-         const emoteSprite = this.emoteSprites[playerId];
-         if (!playerSprite || !emoteSprite || !emoteSymbol) return;
-
-         if (this.emoteTimers[playerId]) clearTimeout(this.emoteTimers[playerId]);
-
-         const emoteAnimKey = `emote_${emoteSymbol}`;
-         if (this.anims.exists(emoteAnimKey)) {
-             emoteSprite.play(emoteAnimKey);
-             emoteSprite.setVisible(true);
-             emoteSprite.setPosition(playerSprite.x, playerSprite.y - playerSprite.displayHeight / 2 - 10);
-
-             // Client handles visual timeout, server clears state later
-             this.emoteTimers[playerId] = setTimeout(() => {
-                 this.hideEmote(playerId);
-             }, 2000);
-         }
-     }
-
-      hideEmote(playerId) {
-          // ... (Implementation similar to previous version) ...
-           const emoteSprite = this.emoteSprites[playerId];
-           if (emoteSprite) emoteSprite.setVisible(false);
-           if (this.emoteTimers[playerId]) {
-               clearTimeout(this.emoteTimers[playerId]);
-               delete this.emoteTimers[playerId];
-           }
-     }
-
-     updateEmotePositions() {
-        for (const playerId in this.emoteSprites) {
-             const emoteSprite = this.emoteSprites[playerId];
-             const playerSprite = this.playerSprites[playerId];
-             if (emoteSprite && playerSprite && emoteSprite.visible) {
-                 emoteSprite.setPosition(playerSprite.x, playerSprite.y - playerSprite.displayHeight / 2 - 10);
-             }
-         }
-     }
-
-     // highlightPlayersInZone and resetPlayerHighlighting would be similar to before
-
-
-} // End of class GameScene
+}
