--- phaser-demo/index.html
@@ -1,26 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Phaser Minimal Demo</title>
-    <style>
-        body {
-            margin: 0;
-            padding: 0;
-            display: flex;
-            justify-content: center;
-            align-items: center;
-            min-height: 100vh;
-            background-color: #222;
-        }
-        #game-container {
-            box-shadow: 0 0 20px rgba(0, 0, 0, 0.5);
-        }
-    </style>
-</head>
-<body>
-    <div id="game-container"></div>
-    <script type="module" src="./src/main.ts"></script>
-</body>
-</html>
