--- phaser-demo/package.json
@@ -1,21 +1,0 @@
-{
-  "name": "phaser-minimal-demo",
-  "version": "1.0.0",
-  "description": "Minimal Phaser demo with Colyseus integration",
-  "scripts": {
-    "start": "parcel index.html",
-    "build": "parcel build index.html"
-  },
-  "author": "",
-  "license": "ISC",
-  "devDependencies": {
-    "buffer": "^6.0.3",
-    "parcel": "^2.16.0",
-    "process": "^0.11.10",
-    "typescript": "^5.9.3"
-  },
-  "dependencies": {
-    "colyseus.js": "^0.16.22",
-    "phaser": "^3.90.0"
-  }
-}
